# projects/01-ai-knowledge-base/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  handler.AskRequestBody:
    type: object
  handler.BatchDeleteRequest:
    properties:
      document_ids:
        items:
          type: string
        maxItems: 100
        minItems: 1
        type: array
    required:
    - document_ids
    type: object
...
  handler.SystemStatusResponse:
    properties:
      runtime:
        $ref: '#/definitions/handler.RuntimeMetrics'
      services:
        additionalProperties:
          $ref: '#/definitions/handler.ServiceInfo'
        type: object
      status:
        type: string
      system:
        $ref: '#/definitions/handler.SystemMetrics'
      timestamp:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
...
  middleware.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorDetail'
      meta:
        $ref: '#/definitions/middleware.MetaInfo'
      success:
        type: boolean
    type: object
...
  middleware.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/middleware.MetaInfo'
      pagination:
        $ref: '#/definitions/middleware.Pagination'
      success:
        type: boolean
    type: object
...
  middleware.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/middleware.MetaInfo'
      success:
        type: boolean
    type: object
...
  service.AskResponse:
    properties:
      answer:
        type: string
      confidence:
        description: 0-1 confidence score
        type: number
      context:
        description: 'Optional: return context used'
        type: string
      kbs_searched:
        type: integer
      latency_ms:
        type: integer
      sources:
        items:
          $ref: '#/definitions/service.Source'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saleschampionhub.com
    name: 技术支持
  description: "企业级AI知识库管理平台，支持多租户、RLS数据隔离和向量搜索功能。\n\n## \U0001F510 认证说明\n所有接口都需要JWT
    Token认证（从子项目0用户中心获取）。\n使用用户中心的 `/api/v1/auth/login` 接口获取Token。\n\n## ✨ 核心特性\n-
    多租户架构：租户/组织/用户三级隔离\n- Row-Level Security (RLS)：数据库级别的数据隔离\n- 三级挂载系统：租户 > 组织 >
    用户权限继承\n- 向量搜索：基于pgvector的语义搜索（1024维向量）\n- Redis多层缓存：热点数据缓存优化\n- 完整用户中心集成：统一认证授权"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AI知识库管理平台 API
  version: "1.0"
paths:
...
  /api/v1/system/status:
    get:
      description: 获取详细的系统状态信息，包括CPU、内存、磁盘、网络等指标
      produces:
      - application/json
      responses:
        "200":
          description: 系统状态信息
          schema:
            $ref: '#/definitions/handler.SystemStatusResponse'
      security:
      - BearerAuth: []
      summary: 获取系统状态
      tags:
      - 监控
  /ask:
    post:
      consumes:
      - application/json
      description: 基于知识库内容回答问题，使用检索增强生成(RAG)技术。通过BGE-large-zh模型(1024维向量)进行语义检索，结合千问(Qwen)大语言模型生成准确回答
      parameters:
      - description: 问答请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.AskRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: 回答成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.AskResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: 无权访问指定知识库
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: RAG 问答
      tags:
      - RAG
  /ask-stream:
    post:
      consumes:
      - application/json
      description: 基于知识库内容回答问题（流式响应），使用检索增强生成(RAG)技术。通过BGE-large-zh模型(1024维向量)进行语义检索，结合千问(Qwen)大语言模型实时流式生成回答
      parameters:
      - description: 问答请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.AskRequestBody'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE流式响应
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: 无权访问指定知识库
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: RAG 问答（流式）
      tags:
      - RAG
  /documents:
    get:
      description: List documents in a knowledge base with pagination and filters
      parameters:
      - description: Knowledge Base ID
        in: query
        name: kb_id
        required: true
        type: string
      - default: 50
        description: Number of results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by status (pending, processing, completed, failed)
        in: query
        name: status
        type: string
      - description: Filter by file type (.pdf, .docx, .md, etc.)
        in: query
        name: file_type
        type: string
      - description: Search in filename and content
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Document'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Upload a document file (PDF, Word, Markdown, TXT, HTML) to a knowledge
        base
      parameters:
      - description: Knowledge Base ID
        in: formData
        name: kb_id
        required: true
        type: string
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      - default: false
        description: Check for duplicate documents
        in: formData
        name: check_duplicate
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "415":
          description: Unsupported file type
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a document to a knowledge base
      tags:
      - documents
  /documents/{id}:
    delete:
      description: Soft delete a document by ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a document
      tags:
      - documents
    get:
      description: Get a document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Document'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a document
      tags:
      - documents
  /documents/{id}/chunks:
    get:
      description: Get all chunks for a document with pagination
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Number of results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.DocumentChunksResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List document chunks
      tags:
      - documents
  /documents/{id}/download:
    get:
      description: Download the original uploaded document file
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Document file
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download document
      tags:
      - documents
  /documents/{id}/preview:
    get:
      description: Get the extracted text content of a document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.DocumentPreviewResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview document content
      tags:
      - documents
  /documents/{id}/status:
    put:
      consumes:
      - application/json
      description: Update a document's status, optionally trigger reprocessing
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDocumentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update document status
      tags:
      - documents
  /documents/batch-delete:
    post:
      consumes:
      - application/json
      description: Delete multiple documents by their IDs
      parameters:
      - description: Batch delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BatchDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.BatchDeleteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch delete documents
      tags:
      - documents
  /documents/batch-update-status:
    post:
      consumes:
      - application/json
      description: Update status for multiple documents at once
      parameters:
      - description: Batch status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BatchUpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.BatchUpdateStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch update document status
      tags:
      - documents
  /embedding:
    post:
      consumes:
      - application/json
      description: 将文本转换为1024维向量，用于语义搜索和RAG问答。使用BGE-large-zh模型
      parameters:
      - description: 文本向量化请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.EmbeddingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 向量生成成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.EmbeddingResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 生成文本向量
      tags:
      - 向量化
  /embeddings/batch:
    post:
      consumes:
      - application/json
      description: 批量将多个文本转换为向量，提高处理效率
      parameters:
      - description: 批量文本向量化请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BatchEmbeddingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量向量生成成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.BatchEmbeddingResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 批量生成文本向量
      tags:
      - 向量化
  /knowledge-bases:
    get:
      description: 分页获取知识库列表，支持多种过滤条件
      parameters:
      - default: 20
        description: 每页数量（默认：20）
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量（默认：0）
        in: query
        name: offset
        type: integer
      - default: created_at DESC
        description: 排序字段（默认：created_at DESC）
        in: query
        name: order_by
        type: string
      - description: 可见性过滤（public/private/restricted）
        in: query
        name: visibility
        type: string
      - description: 所有者ID过滤
        in: query
        name: owner_id
        type: string
      - description: 搜索关键词（在名称和描述中搜索）
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_bases:
                      items:
                        $ref: '#/definitions/models.KnowledgeBase'
                      type: array
                    limit:
                      type: integer
                    offset:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 列出知识库
      tags:
      - 知识库管理
    post:
      consumes:
      - application/json
      description: 创建一个新的知识库。如果未指定所有者，将自动设置为当前用户。
      parameters:
      - description: 知识库创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateKBRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 知识库创建成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_base:
                      $ref: '#/definitions/models.KnowledgeBase'
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建知识库
      tags:
      - 知识库管理
...
  /knowledge-bases/{id}/mounts:
    get:
      description: 获取指定知识库的所有挂载记录，包括租户、组织和用户级别的挂载
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 挂载列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    mounts:
                      items:
                        $ref: '#/definitions/models.KnowledgeBaseMount'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: 知识库ID必填
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取知识库的挂载列表
      tags:
      - 知识库挂载
...
securityDefinitions:
  BearerAuth:
    description: 输入格式："Bearer " + JWT令牌（注意Bearer后面有空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"